backend/requirements.txt
fastapi
uvicorn
sqlmodel
supabase
pydantic
pydantic-settings
python-dotenv
asyncpg
pypdf
PyMuPDF==1.24.0
pdf2image==1.17.0
Pillow
python-multipart
ollama
google-genai
groq
edge-tts
tenacity==8.2.3
tiktoken==0.7.0
pytesseract==0.3.10
pandas==2.1.4
tabulate==0.9.0
paddlepaddle
paddleocr
lxml
